requirement.txt
pyautogui
pyperclip
google-generativeai
keyboard
pyttsx3

so,first install this by:

pip install -r requirements.txt

or also by
pip install pyautogui pyperclip google-generativeai keyboard pyttsx3